requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
atomicwrites==1.4.1
attrs==23.2.0
bcrypt==4.3.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==2.0.12
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
Django==3.2
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
hashids==1.3.1
idna==3.10
iniconfig==2.0.0
packaging==24.0
passlib==1.7.4
pluggy==0.13.1
py==1.11.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2024.1
requests==2.26.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.0
starlette==0.47.2
toml==0.10.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.35.0
